authorization implemented

+++ src/api/songs/handler.js
@@ -16,11 +16,12 @@ class SongsHandler {
     try {
       this._validator.validateSongPayload(request.payload);
       const {
-        title = '', year, performer, genre, duration,
+        title = 'untitled', year, performer, genre, duration,
       } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
       const songId = await this._service.addSong({
-        title, year, performer, genre, duration,
+        title, year, performer, genre, duration, owner: credentialId,
       });
 
       const response = h.response({
@@ -53,8 +54,9 @@ class SongsHandler {
     }
   }
 
-  async getSongsHandler() {
-    const songs = await this._service.getSongs();
+  async getSongsHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const songs = await this._service.getSongs(credentialId);
     return {
       status: 'success',
       data: {
@@ -66,6 +68,9 @@ class SongsHandler {
   async getSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       const song = await this._service.getSongById(id);
       return {
         status: 'success',
@@ -98,7 +103,9 @@ class SongsHandler {
     try {
       this._validator.validateSongPayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.editSongById(id, request.payload);
 
       return {
@@ -129,6 +136,9 @@ class SongsHandler {
   async deleteSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.deleteSongById(id);
 
       return {
